[v0.8.7] refactor: split ApplicationModels.cs into individual entity files
