+++ src/RegistraceOvcina.Web/Data/Models/ExternalContact.cs
@@ -1,0 +1,8 @@
+namespace RegistraceOvcina.Web.Data;
+
+public sealed class ExternalContact
+{
+    public int Id { get; set; }
+    public string Email { get; set; } = "";
+    public DateTime CreatedAtUtc { get; set; }
+}
