+++ src/RegistraceOvcina.Web/Data/Models/Room.cs
@@ -1,0 +1,8 @@
+namespace RegistraceOvcina.Web.Data;
+
+public sealed class Room
+{
+    public int Id { get; set; }
+    public string Name { get; set; } = "";
+    public int DefaultCapacity { get; set; }
+}
